modifying project structure and handling rating in watchlist page
