+++ src/App.jsx
@@ -17,11 +17,7 @@ const CategoryPage = lazy(() => import("./pages/CategoryPage"));
 const Movies = lazy(() => import("./pages/MoviesPage"));
 
 function App() {
-	const wishItem = useSelector((state) => state.wishlist.wishItem || []);
 
-	useEffect(() => {
-		localStorage.setItem("wishList", JSON.stringify(wishItem));
-	}, [wishItem]);
 
 	return (
 		<BrowserRouter>
